Refactoring. Interactive selection of components.
